+++ src/x2net/Coroutine.cs
@@ -46,7 +46,8 @@ namespace x2net
     {
         Ok,
         Error,
-        Timeout
+        Timeout,
+        Canceled,
     }
 
     /// <summary>